+++ compile.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-gcc x4.c -o x4 -lX11 -lm
+gcc x4.c configuracion.c -o x4 -lX11 -lm