+++ changelog.md
@@ -2,6 +2,18 @@
 
 ## Versiones
 
+### Versión 1.0.3 - Junio 2021
+- Añadido un fichero de configuración x4.conf
+- Opción de leer del fichero de configuración los aspectos básicos:
+  - Ancho del borde de las herramientas
+  - Tamaño de la flecha
+  - Fuente
+    - La fuente por defecto es -*-*-*-*-*-*-60-*-*-*-*-*-iso8859-*
+    - Se puede saber las fuentes disponibles en el sistema con la orden xlsfont o xfontsel -scaled
+    - No se pueden usar fuentes TTF
+  - Número de niveles de deshacer
+  
+
 ### Versión 1.0.2 - Junio 2021
 - Añadida la función de dibujar números que indican "pasos" de un procedimiento